+++ spans/__init__.py
@@ -11,7 +11,7 @@ together.
 
 """
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 __all__ = [
 	"intrange",
 	"floatrange",